+++ src/main/java/dev/fulmineo/companion_bats/CompanionBatsClient.java
@@ -7,12 +7,14 @@ import net.fabricmc.api.Environment;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.fabric.api.client.rendereregistry.v1.EntityRendererRegistry;
 import net.fabricmc.fabric.api.client.screenhandler.v1.ScreenRegistry;
+import net.minecraft.client.render.entity.FlyingItemEntityRenderer;
 
 @Environment(EnvType.CLIENT)
 public class CompanionBatsClient implements ClientModInitializer {
     @Override
     public void onInitializeClient() {
         ScreenRegistry.register(CompanionBats.BAT_SCREEN_HANDLER, CompanionBatScreen::new);
-        EntityRendererRegistry.INSTANCE.register(CompanionBats.COMPANION_BAT, (dispatcher) -> new CompanionBatEntityRenderer(dispatcher));
+        EntityRendererRegistry.INSTANCE.register(CompanionBats.COMPANION_BAT, CompanionBatEntityRenderer::new);
+        EntityRendererRegistry.INSTANCE.register(CompanionBats.DYNAMITE, FlyingItemEntityRenderer::new);
     }
 }