+++ src/main/java/dev/fulmineo/companion_bats/entity/ai/goal/CompanionBatPickUpItemGoal.java
@@ -50,7 +50,7 @@ public class CompanionBatPickUpItemGoal extends Goal {
             } else if (this.isBundleAvailable()) {
                 this.owner = livingEntity;
 				if (this.owner.handSwinging) return false;
-                this.itemList = this.entity.world.getEntitiesByClass(ItemEntity.class, this.entity.getBoundingBox().expand(12.0D), (itemEntity) -> true);
+                this.itemList = this.entity.world.getEntitiesByClass(ItemEntity.class, this.entity.getBoundingBox().expand(12.0D), (itemEntity) -> itemEntity.getThrower() != this.owner.getUuid());
                 this.itemList.sort(new ProximityComparator());
                 return this.itemList.size() > 0;
             }